requirements.txt
pyzmq>=25.0.0
pyserial>=3.5
